Get user orders by user email instead of rely on current user
The problem was because `is_user_logged_in` it was not working 100% all the
times and the:

- POST
- GET

Endpoints for `create-order` was not working as expected, in this case a change
to use the user email is a more consistent value same as it's working with the
`token_id`.

This change also fix a bug where with any random token we were able to get all
the orders from all the anonymous users.